+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Display "Follow" button to guests.
+
 ### Changed
 
 - Make multiple requests when refreshing follower status.